Bug fix in FMM_Pts, FMM_Tree
Vector for [up/dn]wd_[equiv/check]_surface coordinates was a member of
FMM_Pts, however the vector points to memory in setup_data which belongs
to FMM_Tree.  This caused access to freed memory if the tree was
destroyed.  The fix is to move this coordinate data to the tree.