Fix projection for periodic solutions
For periodic solutions of Poisson and Stokes equations, the
compatibility condition requires that the sum (or integral) of the
forcing term must be zero. For problems where this is not satisfied, we
still try to obtain a periodic solution by adding a uniform volume
force. This requires volume integrals and for particle FMM. For particle
FMM, we fix this without having to compute such integrals.

- For the multipole expansion of the root node, we precompute the
  integral of the kernel function over the unit cube. This fixes the
  issue for all octants well-separated from the root octant.

- The remaining integral is over [-1,2]^3. To add this contribution,
  without actually computing the integral, the user provides a
  particular solution of the PDE Lu=1 in this region. The difference
  between the particular solution and the required integral is added
  to the local expansion of the root node. Then, at each target point,
  we evaluate the particular solution and add it to the potential at the
  point.